// ui-app/src/app/api/models/user-details-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserDetailsResponseDto {
  admin?: boolean;
  email?: string;
  id?: number;
  roles?: Array<string>;
  username?: string;
  vigyanKendraCode?: string;
  vigyanKendraId?: number;
  vigyanKendraName?: string;
}